core: Move as3lint to build_playerglobal
