--- as3lint/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "as3lint"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-serde = {version = "1.0.143", features = ["derive"]}
-serde-xml-rs = "0.5.1"
-colored = "2.0.0"
