+++ core/build_playerglobal/Cargo.toml
@@ -8,3 +8,7 @@ convert_case = "0.5.0"
 proc-macro2 = "1.0.43"
 quote = "1.0.21"
 swf = { path = "../../swf" }
+clap = {version = "3.2.17", features = ["derive"]}
+serde = {version = "1.0.143", features = ["derive"]}
+serde-xml-rs = "0.5.1"
+colored = "2.0.0"
